requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.27
langchain-core==0.3.76
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
langsmith==0.4.27
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.0.2
orjson==3.11.3
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.8
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
pyzmq==27.1.0
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.14
zipp==3.23.0
zstandard==0.24.0
